Fix ExplicitMux dispatch: 1-byte path size, ForwardOpen no-fallthrough, accept logs
Three fixes to the ExplicitHandler hook path, surfaced during
end-to-end testing against a Studio 5000 PLC:

  1. unconnectedExplicitMux read 2 bytes for path size; CIP Vol 1
     §3-4.4.1 specifies one byte (count of 16-bit words). The wrong
     read consumed the first EPATH byte as the high byte of the size,
     producing absurd "path size 8196 words extends past item end"
     errors on every Studio 5000 unconnected MSG. Fix: read pathSize
     as one byte. Test rewritten to match the real wire shape.

  2. unconnectedData's switch handles Forward_Open, LargeForwardOpen,
     and Forward_Close — each handler reads the request and sends
     its own reply. But the case arms didn't return: control fell
     through to the post-switch ExplicitMux call, which read garbage
     as path-size and emitted a duplicate reply. Studio 5000 cached-
     connected MSGs (ConnectedFlag=2) silently failed this way. Fix:
     explicit `return nil` after each Forward_* handler block.

  3. Added INFO-level logs at TCP accept, EIP frame dispatch, and
     unhandled EIP commands. Useful for diagnosing PLC↔server
     connectivity without packet capture.

+++ server_mux.go
@@ -261,12 +261,13 @@ func (h *serverTCPHandler) connectedExplicitMux(service CIPService, items []CIPI
 // unconnectedExplicitMux is the unconnected-dispatch entry point invoked from
 // unconnectedData when the service is Set_Attribute_Single or any other
 // service that falls through to the explicit mux. The caller has already
-// consumed the service byte; item is positioned at the path-size word.
+// consumed the service byte; item is positioned at the path-size byte.
 //
-// Unconnected explicit messages carry: path-size word, EPATH, then optional
-// service data.
+// Unconnected explicit messages carry: path-size byte (in 16-bit words),
+// EPATH, then optional service data — per CIP Vol 1 §3-4.4.1. There is no
+// pad byte between the size and the EPATH.
 func (h *serverTCPHandler) unconnectedExplicitMux(service CIPService, item *CIPItem) error {
-	pathSize, err := item.Uint16()
+	pathSize, err := item.Byte()
 	if err != nil {
 		return fmt.Errorf("read path size: %w", err)
 	}
